Split film-details into components. Part 2
