+++ src/components/button-show-more/button-show-more.js
@@ -1,5 +1,5 @@
 /**
- * Создание шаблона кнопки показа оставшихся фильмов
- * @return {string} кнопка
+ * Создание разметки кнопки показа фильмов
+ * @return {string} разметка кнопки показа скрытых фильмов
  */
 export const createShowMore = () => `<button class="films-list__show-more">Show more</button>`;
