+++ src/components/films/components/details-info.js
@@ -1,0 +1,68 @@
+import {createGenresList} from "./details-info/genre-list";
+
+
+/**
+ * Создание разметки блока подробной информации о фильме
+ * @param {*} film фильм
+ * @return {string} разметка блока информации
+ */
+export const createDetailsInfo = (film) => {
+  const {poster, title, titleOrigin, rating, director, screenwriters,
+    actors, releaseDate, duration, country, genres, description, age} = film;
+
+  return (`
+    <div class="film-details__info-wrap">
+      <div class="film-details__poster">
+        <img class="film-details__poster-img" src="./images/posters/${poster}" alt="">
+
+        <p class="film-details__age">${age}</p>
+      </div>
+
+      <div class="film-details__info">
+        <div class="film-details__info-head">
+          <div class="film-details__title-wrap">
+            <h3 class="film-details__title">${title}</h3>
+            <p class="film-details__title-original">Original: ${titleOrigin}</p>
+          </div>
+
+          <div class="film-details__rating">
+            <p class="film-details__total-rating">${rating}</p>
+          </div>
+        </div>
+
+        <table class="film-details__table">
+          <tr class="film-details__row">
+            <td class="film-details__term">Director</td>
+            <td class="film-details__cell">${director}</td>
+          </tr>
+          <tr class="film-details__row">
+            <td class="film-details__term">Writers</td>
+            <td class="film-details__cell">${screenwriters}</td>
+          </tr>
+          <tr class="film-details__row">
+            <td class="film-details__term">Actors</td>
+            <td class="film-details__cell">${actors}</td>
+          </tr>
+          <tr class="film-details__row">
+            <td class="film-details__term">Release Date</td>
+            <td class="film-details__cell">${releaseDate}</td>
+          </tr>
+          <tr class="film-details__row">
+            <td class="film-details__term">Runtime</td>
+            <td class="film-details__cell">${duration}</td>
+          </tr>
+          <tr class="film-details__row">
+            <td class="film-details__term">Country</td>
+            <td class="film-details__cell">${country}</td>
+          </tr>
+          ${createGenresList(genres)}
+        </table>
+
+        <p class="film-details__film-description">
+          ${description}
+        </p>
+      </div>
+    </div>
+
+  `);
+};
