+++ src/components/films/components/details-info/genre-item.js
@@ -1,0 +1,10 @@
+/**
+ * Создание разметки жанра
+ * @param {string} genre жанр
+ * @return {string} разметка элемента
+ */
+export const createGenre = (genre) => {
+  return (`
+    <span class="film-details__genre">${genre}</span>
+  `);
+};
