+++ src/components/films/film-details.js
@@ -1,36 +1,14 @@
-import {createCommentList} from "./components/comment-list";
+import {createDetailsInfo} from "./components/details-info";
 import {createControls} from "./components/controls";
-import {CommentEmojies} from "../../consts";
-import {createEmojiList} from "./components/emoji-list";
+import {createCommentBlock} from "./components/comments";
+
 
 /**
- * Создание разметки подробной карточки о фильме
+ * Создание разметки подробной карточки фильма
  * @param {Object} film фильм
- * @return {string} разметка подробной карточки о фильме
+ * @return {string} разметка подробной карточки фильма
  */
 export const createFilmDetails = (film) => {
-  const {
-    poster,
-    title,
-    titleOrigin,
-    rating,
-    director,
-    screenwriters,
-    actors,
-    releaseDate,
-    year,
-    duration,
-    country,
-    genres,
-    commentsCount,
-    comments,
-    description,
-    age,
-    isWatch,
-    isWatched,
-    isFavorite
-  } = film;
-
   return (`
     <section class="film-details">
       <form class="film-details__inner" action="" method="get">
@@ -38,85 +16,10 @@ export const createFilmDetails = (film) => {
           <div class="film-details__close">
             <button class="film-details__close-btn" type="button">close</button>
           </div>
-          <div class="film-details__info-wrap">
-            <div class="film-details__poster">
-              <img class="film-details__poster-img" src="./images/posters/${poster}" alt="">
-
-              <p class="film-details__age">${age}</p>
-            </div>
-
-            <div class="film-details__info">
-              <div class="film-details__info-head">
-                <div class="film-details__title-wrap">
-                  <h3 class="film-details__title">${title}</h3>
-                  <p class="film-details__title-original">Original: ${titleOrigin}</p>
-                </div>
-
-                <div class="film-details__rating">
-                  <p class="film-details__total-rating">${rating}</p>
-                </div>
-              </div>
-
-              <table class="film-details__table">
-                <tr class="film-details__row">
-                  <td class="film-details__term">Director</td>
-                  <td class="film-details__cell">${director}</td>
-                </tr>
-                <tr class="film-details__row">
-                  <td class="film-details__term">Writers</td>
-                  <td class="film-details__cell">${screenwriters}</td>
-                </tr>
-                <tr class="film-details__row">
-                  <td class="film-details__term">Actors</td>
-                  <td class="film-details__cell">${actors}</td>
-                </tr>
-                <tr class="film-details__row">
-                  <td class="film-details__term">Release Date</td>
-                  <td class="film-details__cell">${releaseDate}</td>
-                </tr>
-                <tr class="film-details__row">
-                  <td class="film-details__term">Runtime</td>
-                  <td class="film-details__cell">${duration}</td>
-                </tr>
-                <tr class="film-details__row">
-                  <td class="film-details__term">Country</td>
-                  <td class="film-details__cell">${country}</td>
-                </tr>
-                <tr class="film-details__row">
-                  <td class="film-details__term">Genres</td>
-                  <td class="film-details__cell">
-                    <span class="film-details__genre">Drama</span>
-                    <span class="film-details__genre">Film-Noir</span>
-                    <span class="film-details__genre">Mystery</span></td>
-                </tr>
-              </table>
-
-              <p class="film-details__film-description">
-                ${description}
-              </p>
-            </div>
-          </div>
-
-          ${createControls()}
+          ${createDetailsInfo(film)}
+          ${createControls(film)}
         </div>
-
-        <div class="form-details__bottom-container">
-          <section class="film-details__comments-wrap">
-            <h3 class="film-details__comments-title">Comments <span class="film-details__comments-count">${commentsCount}</span></h3>
-
-            ${createCommentList(comments)}
-
-            <div class="film-details__new-comment">
-              <div for="add-emoji" class="film-details__add-emoji-label"></div>
-
-              <label class="film-details__comment-label">
-                <textarea class="film-details__comment-input" placeholder="Select reaction below and write comment here" name="comment"></textarea>
-              </label>
-
-              ${createEmojiList(CommentEmojies)}
-            </div>
-          </section>
-        </div>
+        ${createCommentBlock(film)}
       </form>
     </section>
   `);
